refactor(chanx): return int64 for queue lengths
Use atomic.Int64 for unbounded queue length tracking and expose
length snapshots as int64 across chanx and chanrpc.

BREAKING CHANGE: chanx Unbounded Len and BufLen, and chanrpc Client and Server Len, now return int64 instead of int

+++ chanrpc/client.go
@@ -41,7 +41,7 @@ func (c *Client) Event() <-chan *RetInfo {
 	return c.chanAsyncRet.Out()
 }
 
-func (c *Client) Len() int {
+func (c *Client) Len() int64 {
 	return c.chanAsyncRet.Len()
 }
